refactor: move context listeners to their own package
